refactor: util/format에 게시물, 댓글의 작성 시간의 형식을 반환하는 getTimeDifference 메서드 추가 및 중복 코드 삭제
