+++ src/app/(causw)/setting/my/comments/page.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import { useRouter } from "next/navigation";
+import { Icon, PreviousButton, SettingService } from "@/shared";
 import { useEffect, useRef, useState } from "react";
 
 import Image from "next/image";
-
 import { LoadingComponent } from "@/entities";
-
-import { Icon, PreviousButton, SettingService } from "@/shared";
+import { getTimeDifference } from "@/utils/format";
+import { useRouter } from "next/navigation";
 
 const MyCommentsPage = () => {
   const { getMyCommentPosts } = SettingService();
@@ -150,28 +149,6 @@ const MyCommentsPage = () => {
       )}
     </div>
   );
-};
-
-const getTimeDifference = (ISOtime: string) => {
-  const createdTime = new Date(ISOtime);
-  const now = new Date();
-  const diffMSec = now.getTime() - createdTime.getTime();
-  const diffMin = Math.round(diffMSec / (60 * 1000));
-  if (diffMin === 0) {
-    return `방금 전`;
-  } else if (diffMin < 60) {
-    return `${diffMin}분 전`;
-  } else if (
-    now.getFullYear() === createdTime.getFullYear() &&
-    now.getMonth() === createdTime.getMonth() &&
-    now.getDate() === createdTime.getDate()
-  ) {
-    return `${createdTime.getHours()}:${createdTime.getMinutes()}`;
-  } else if (now.getFullYear() === createdTime.getFullYear()) {
-    return `${createdTime.getMonth() + 1}/${createdTime.getDate()}`;
-  } else {
-    return `${now.getFullYear() - createdTime.getFullYear()}년 전`;
-  }
 };
 
 export default MyCommentsPage;
